fix(relayer): cast size/cap params to bigint in the atomic upsert
The payment-idempotency atomic upsert (#350) used $4 (size) both as the
bigint size column and in `capacity.staged + $4` where staged is
numeric (SUM), so Postgres could not deduce a single type and every
ingest failed at runtime with 'inconsistent types deduced for parameter
$4' (numeric versus bigint). FakePool unit tests do not exercise real
type inference, so this passed CI and only surfaced against live
Postgres. Cast $4 and $7 to bigint. Verified with PREPARE against the
live schema: the old form errors, the cast form prepares clean.

+++ relayer/src/staged/staged-intent.store.ts
@@ -162,9 +162,9 @@ export class StagedIntentStore {
        )
        INSERT INTO ${TABLE}
          (intent_id, bytes, blob_id, size, next_attempt_at, created_at, updated_at)
-       SELECT $1, $2, $3, $4, $5, $5, $5
+       SELECT $1, $2, $3, $4::bigint, $5, $5, $5
          FROM capacity
-        WHERE $6 = false OR capacity.staged + $4 <= $7
+        WHERE $6 = false OR capacity.staged + $4::bigint <= $7::bigint
        ON CONFLICT (intent_id) DO UPDATE SET
          bytes = EXCLUDED.bytes,
          blob_id = EXCLUDED.blob_id,
@@ -172,7 +172,7 @@ export class StagedIntentStore {
          updated_at = EXCLUDED.updated_at
        WHERE ${TABLE}.state = 'active'
          AND ($6 = false
-              OR (SELECT staged FROM capacity) + EXCLUDED.size <= $7)
+              OR (SELECT staged FROM capacity) + EXCLUDED.size <= $7::bigint)
        RETURNING intent_id`,
       [intentId, blob.bytes, blob.blobId, blob.size, now, capped, cap],
     );
